docs: 디렉토리 통합2

+++ back/src/controllers/challengeController.ts
@@ -6,11 +6,11 @@ import {
   updateChallenge,
   deleteChallenge,
 } from '@_services/challengeService';
+import { User } from '@_types/user'
 
-// 모든 챌린지 가져오기
 export const getAllChallengesController = async (req: Request, res: Response) => {
+  const searchQuery = req.query.search ? req.query.search.toString() : '';
   try {
-    const searchQuery = req.query.search ? req.query.search.toString() : '';
     const challenges = await getAllChallenges(searchQuery);
     res.status(200).json({ challenges });
   } catch (error) {
@@ -19,7 +19,6 @@ export const getAllChallengesController = async (req: Request, res: Response) =>
   }
 };
 
-// 특정 챌린지 가져오기
 export const getChallengeByIdController = async (req: Request, res: Response) => {
   const { id } = req.params;
   try {
@@ -31,37 +30,50 @@ export const getChallengeByIdController = async (req: Request, res: Response) =>
   }
 };
 
-// 챌린지 생성하기
 export const createChallengeController = async (req: Request, res: Response) => {
   const { title, description, start_date, end_date, tasks } = req.body;
-  try {
-    const newChallenge = await createChallenge({ title, description, start_date, end_date, tasks});
+  try { 
+    const userId = (req.user as User).id; 
+    if (!userId) {
+      return res.status(401);
+    }
+    const newChallenge = await createChallenge(userId, { title, description, start_date, end_date, tasks });
     res.status(201).json(newChallenge);
+
   } catch (error) {
     console.error('챌린지 생성을 실패 했습니다.:', error);
     res.status(500).send('서버 오류발생');
+
   }
 };
 
-// 챌린지 수정하기
 export const updateChallengeController = async (req: Request, res: Response) => {
   const { id } = req.params;
   const { title, description, start_date, end_date } = req.body;
   try {
+    const { challenge } = await getChallengeById(id);
+    if (challenge.user_id !== (req.user as User).id) {  
+      return res.status(403);
+    }
     const updatedChallenge = await updateChallenge(id, { title, description, start_date, end_date });
     res.status(204).json(updatedChallenge);
+    
   } catch (error) {
     console.error(`${id} 아이디의 챌린지 수정을 실패 했습니다.:`, error);
     res.status(500).send('서버 오류발생');
   }
 };
 
-// 챌린지 삭제하기
 export const deleteChallengeController = async (req: Request, res: Response) => {
   const { id } = req.params;
-  try {
+  try { 
+    const { challenge } = await getChallengeById(id);
+    if (challenge.user_id !== (req.user as User).id) {  // Optional Chaining 사용
+      return res.status(403);
+    }
     await deleteChallenge(id);
     res.status(204).send();
+    
   } catch (error) {
     console.error(`${id} 아이디의 챌린지 삭제를 실패 했습니다.:`, error);
     res.status(500).send('서버 오류발생');
